Updates to clx streamz structure
